moved Extended.cs to folder
